move devnet utilities in a separate folder in common
