+++ pkg/common/devnet/constants.go
@@ -1,7 +1,8 @@
-package common
+package devnet
 
 // Foundry Image Date : 21 April 2025
 const FOUNDRY_IMAGE = "ghcr.io/foundry-rs/foundry:nightly-1ae64e38a1c69bda45343947875f7c86bad00038"
 const CHAIN_ARGS = "--block-time 3 --base-fee 0 --gas-price 0"
 const FUND_VALUE = "10000000000000000000"
 const RPC_URL = "http://localhost:8545"
+const DEVNET_ENV_KEY = "devnet"
